--- dreamy-backend/server.js
@@ -1,180 +1,0 @@
-const path = require('path');
-const express = require('express');
-const app = express();
-const mongoose = require('mongoose');
-const bodyParser = require('body-parser');
-
-const User = require('./src/models/user');
-const Entry = require('./src/models/entry');
-
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
-
-mongoose.connect('mongodb://n8:nstn8e81@ds259105.mlab.com:59105/dreamers', {useMongoClient: true})
-var port = 5000;
-
-var router = express.Router();
-
-//middleware
-router.use(function(req, res, next){
-  console.log("something is happening");
-  next();
-})
-
-app.use('/api', router);
-// works //
-router.route('/users')
-  .get(function(req, res) {
-    User.find(function(err, users){
-      if (err)
-      console.log(users)
-        res.json(users);
-    });
-  })
-//works//
-router.route('/user/create')
-  .post(function(req, res) {
-    var user = new User();
-    user._id = mongoose.Types.ObjectId();
-    user.dateCreated = new Date();
-    user.firstName = req.body.firstName;
-    user.lastName = req.body.lastName;
-    user.email = req.body.email;
-    user.birthDate = req.body.birthDate;
-    user.age = req.body.age;
-    user.sign = req.body.sign;
-    user.gender = req.body.gender;
-    user.interests = req.body.interests;
-    user.bio = req.body.bio;
-    user.save(function(err){
-      if(err)
-        res.send(err);
-      res.json({message: "User created"});
-    });
-  })
-//works//
-router.route('/user/find/:userid')
-  .get(function(req, res) {
-    console.log(req.params.userid);
-    User.findById(req.params.userid, function(err, user){
-      if (err)
-        res.send(err);
-        console.log(user);
-      res.json(user);
-    });
-  })
-//works//
-router.route('/user/update/:userid')
-  .post(function(req, res) {
-    console.log("update user")
-    User.findById(req.params.userid, function(err, user){
-      if (err)
-        res.send(err);
-        user.firstName = req.body.firstName ? req.body.firstName:user.firstName;
-        user.lastName = req.body.lastName ? req.body.lastName:user.lastName;
-        user.email = req.body.email ? req.body.email:user.email;
-        user.birthDate = req.body.birthDate ? req.body.birthDate:user.birthDate;
-        user.age = req.body.age ? req.body.age:user.age;
-        user.sign = req.body.sign ? req.body.sign:user.sign;
-        user.gender = req.body.gender ? req.body.gender:user.gender;
-        user.interests = req.body.interests ? req.body.interests:user.interests;
-        user.bio = req.body.bio ? req.body.bio:user.bio;
-
-      user.save(function(err){
-        if(err)
-          res.send(err);
-          res.json({message: "User updated"});
-      });
-    });
-  })
-//works to delete user but not journal entries//
-router.route('/user/delete/:userid')
-  .delete(function(req, res) {
-    User.remove({
-      _id: req.params.userid
-    }, function(err, user){
-      if (err)
-        res.send(err);
-
-      res.json({ message: 'User deleted'});
-    });
-  });
-//works//
-router.route('/journals')
-  .get(function(req, res) {
-      Entry.find(function(err, entry){
-        if (err)
-          res.send(err);
-        res.json(entry);
-      });
-  })
-//works//
-router.route('/journal/:journalid')
-  .get(function(req, res) {
-      Entry.findById(req.params.journalid, function(err, journal){
-        if (err)
-          res.send(err);
-        res.json(journal)
-      });
-  });
-// seems to work but does not actually delete from mlab//
-router.route('/journal/delete/:journalid')
-  .delete(function(req, res) {
-      Entry.remove({
-        id: req.params.journalid
-      }, function(err, journal){
-        if (err)
-          res.send(err);
-
-        res.json({ message: 'Entry Deleted'});
-      });
-  });
-
-  // "journalEntries": [
-  //         "5a11fa6a316a410d617ccf57",
-  //         "5a12063439492f11b3ccb0d6",
-  //         "5a12070928478f11d632dab6"
-  //     ]
-// ? //
-router.route('/user/journals/:userid')
-  .get(function(req, res) {
-    User.findById(req.params.userid)
-        .populate("journalEntries")
-        .exec(function (err, entry) {
-          if (err)
-            res.send(err);
-          console.log('The entry id is: %s', entry._id);
-        });
-  })
-
-router.route('/journals/:userid')
-  .post(function(req, res) {
-    User.findById(req.params.userid, function(err, user){
-      if (err)
-        res.send(err);
-        var entry = new Entry();
-        entry._id = mongoose.Types.ObjectId();
-        entry.dateCreated = new Date();
-        entry.createdBy = user._id;
-        entry.dreamDate = req.body.dreamDate;
-        entry.entryTitle = req.body.entryTitle;
-        entry.anonymous = req.body.anonymous;
-        entry.private = req.body.private;
-        entry.description = req.body.description;
-        entry.personalNotes = req.body.personalNotes;
-        entry.bedTime = req.body.bedTime;
-        entry.wakeTime = req.body.wakeTime;
-        user.journalEntries.push(entry);
-        entry.save(function(err, entry){
-          if(err)
-            res.send(err);
-        })
-        user.save(function(err, user){
-          if(err)
-            res.send(err);
-          res.json(user)
-        });
-    });
-  })
-
-app.listen(port);
